Get bootstrap working
Brought over original home page and installed bootstrap 3. This required creating partials for the home page so I could see what the file was doing (original was over 140 lines).

There are significant changes in behavior between v1 and v3 of
bootstrap, so layout work will need to be redone, but that will be a
later task.